# docs/swagger.yaml
basePath: /
definitions:
  dto.GenericResponseDTO:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: FGA Digitalent Final Submission
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MyGram API
  version: "1.0"
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      description: Login with existing user account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.GenericResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.GenericResponseDTO'
      summary: Login User
      tags:
      - user
  /user/register:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      description: Register A User Account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GenericResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.GenericResponseDTO'
      summary: Register A User Account
      tags:
      - user
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"